JCF: Issue #171: move the daq_application-specific headers from apps/ to src/
